+++ aqua/diagnostics/__init__.py
@@ -19,7 +19,7 @@ from .ocean_stratification import Stratification, PlotStratification, PlotMLD
 from .ocean_trends import Trends, PlotTrends
 
 
-DIAGNOSTIC_CONFIG_DIRECTORIES = ["analysis", "tools", "collections"]
+DIAGNOSTIC_CONFIG_DIRECTORIES = ["analysis", "tools", "collections", "definitions"]
 DIAGNOSTIC_TEMPLATE_DIRECTORIES = ["collections"]
 
 __all__ = ["__version__", "NAO", "ENSO", "MJO",