+++ aqua/diagnostics/lat_lon_profiles/lat_lon_profiles.py
@@ -82,7 +82,6 @@ class LatLonProfiles(Diagnostic):
         # Set the region based on the region name or the lon and lat limits
         self.region, self.lon_limits, self.lat_limits = self._set_region(
             region=region,
-            diagnostic="lat_lon_profiles",
             regions_file_path=regions_file_path,
             lon_limits=lon_limits,
             lat_limits=lat_limits,