+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 Unreleased in the current development version (target v0.24.0):
 
 Complete list:
+- Regions config file centralization (#267)
+- Jinja templates for configuration files for collections (#230)
 - Fallback test download from wilma (#269)
 - Jinja templates for configuration files for collections (#230, #270)
 - Histogram: diagnostic updates (#255)